+++ CHANGELOG.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - The `--latest` flag in the `search` command, to use the latest version.
 - Progress bars are now shown while build sources or prebuilds are being downloaded.
 - The `--target-only` flag on the `search` command, to show only the packages that are supported for the current target.
+- The `config repositories set-url` command, which can be used to change the url of a repository.
+- The `config repositories set-prebuilds` command, which can be used to change the prebuilds url of a repository.
+- The `config repositories disable-prebuilds` command, which can be used to enable or disable prebuilds of a repository.
+- The `config repositories add` command now has a `--unchecked` flag, which can be used to skip repository checks.
 
 ### Changes
 - Change display of true and false values in the `info` and `search` commands.
@@ -29,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - The `prebuild_url` and `prebuild_provider` fields in the `repository.toml` file now represent a suggestion that needs to be added to the config in order to be used.
 - The `Alterations` verifier check is now enabled for packages that were installed from a prebuild.
 - The `search --regex` results are now sorted by package name.
+- The `config show` command now also shows the `prebuilds_disabled` option.
 
 ### Fixed
 - Fix error messages not representing the error correctly.